+++ engine/schema/src/main/java/com/cloud/upgrade/dao/Upgrade41520to41600.java
@@ -33,8 +33,6 @@ import org.apache.cloudstack.acl.RoleType;
 import org.apache.log4j.Logger;
 
 import com.cloud.utils.exception.CloudRuntimeException;
-import java.math.BigInteger;
-import java.util.UUID;
 
 
 public class Upgrade41520to41600 implements DbUpgrade, DbUpgradeSystemVmTemplate {
@@ -78,29 +76,6 @@ public class Upgrade41520to41600 implements DbUpgrade, DbUpgradeSystemVmTemplate
         generateUuidForExistingSshKeyPairs(conn);
         populateAnnotationPermissions(conn);
         correctGuestOsIdsInHypervisorMapping(conn);
-        fixWrongPoolUuid(conn);
-    }
-
-    public void fixWrongPoolUuid(Connection conn) {
-        LOG.debug("Replacement of faulty pool uuids");
-        try (PreparedStatement pstmt = conn.prepareStatement("SELECT id,uuid FROM storage_pool "
-                + "WHERE uuid NOT LIKE \"%-%-%-%\" AND removed IS NULL;"); ResultSet rs = pstmt.executeQuery()) {
-            PreparedStatement updateStmt = conn.prepareStatement("update storage_pool set uuid = ? where id = ?");
-            while (rs.next()) {
-                    UUID poolUuid = new UUID(
-                            new BigInteger(rs.getString(2).substring(0, 16), 16).longValue(),
-                            new BigInteger(rs.getString(2).substring(16), 16).longValue()
-                    );
-                    updateStmt.setLong(2, rs.getLong(1));
-                    updateStmt.setString(1, poolUuid.toString());
-                    updateStmt.addBatch();
-            }
-            updateStmt.executeBatch();
-        } catch (SQLException ex) {
-            String errorMsg = "fixWrongPoolUuid:Exception while updating faulty pool uuids";
-            LOG.error(errorMsg,ex);
-            throw new CloudRuntimeException(errorMsg, ex);
-        }
     }
 
     private void correctGuestOsIdsInHypervisorMapping(final Connection conn) {
